+++ labgrid/resource/remote.py
@@ -217,3 +217,15 @@ class NetworkDeditecRelais8(RemoteUSBResource):
     def __attrs_post_init__(self):
         self.timeout = 10.0
         super().__attrs_post_init__()
+
+
+@target_factory.reg_resource
+@attr.s(cmp=False)
+class NetworkSysfsGPIO(NetworkResource, ManagedResource):
+    manager_cls = RemotePlaceManager
+
+    """The NetworkSysfsGPIO describes a remotely accessible gpio line"""
+    index = attr.ib(validator=attr.validators.optional(attr.validators.instance_of(int)))
+    def __attrs_post_init__(self):
+        self.timeout = 10.0
+        super().__attrs_post_init__()